nouncil-recorder/requirements.txt
discord.py==2.3.2
PyAudio==0.2.14
notion-client==2.1.0
python-dotenv==1.0.0